requirements.txt
Brlapi==0.6.1
Django==1.9.5
Pillow==3.2.0
PyICU==1.5
apt-clone==0.2.1
chardet==2.2.1
colorama==0.2.5
command-not-found==0.3
defer==1.0.6
dirspec==13.10
fbmessenger==0.2.0
friends==0.1
html5lib==0.999
httplib2==0.8
language-selector==0.1
louis==2.5.3
lxml==3.3.3
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
piston-mini-client==0.7.5
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
requests==2.2.1
six==1.5.2
thin-client-config-agent==0.8
ubuntu-drivers-common==0.0.0
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
urllib3==1.7.1
virtkey==0.63.0
virtualenv==15.0.1
xkit==0.0.0
